requirements-full.txt
# Full requirements for advanced quant, ML, and deep learning
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pytz>=2023.3
pathlib2>=2.3.7
yfinance>=0.2.0
ta>=0.10.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.2
tensorflow==2.15.0  # Uncommented for Docker build with all system dependencies
keras>=2.12.0
torch>=2.0.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
redis>=4.6.0
loguru>=0.7.0

